Merged PR 9875: Resolve "Request Entity Too Large" Error When Updating Cases
The close cases service has been receiving an http 413 - Content Too Large response when closing incidents in batch. This turns out to be due to an error in the query construction.

Previously the service attempted to update hundreds of incidents in batch by constructing huge transactions (250 rows each). Since each incident is independent of the others, this PR replaces these huge transactions with smaller ones - each transaction now consists of a pair of one update request and one close incident request. As these should probably be updated together (and fail together) it makes sense to have them in a transaction.

All of these "smaller" transactions are then send in an execute multiple request, which seems to have fixed the 413 problem (tested against TEST and ACC).

The new execute multiple requests give rise to a timeout error, however, so this is resolved by reducing the batch size to 100. One batch completes in about 40 seconds in TEST and ACC, but this can of course be tweaked later. This should allow the job to clean up production (ca 16,000 open incidents to close) in about 2-3 hours.

# Change Log

## Close Cases Service
- Switch to creating pairs of small transactions for each incident to be updated
- Remove unnecessary columns from the query if we are in prodcution mode

## Helper Class:
- Switch from Boolean to bool
- Add trace log for the batch request
- Add handling of timeout exceptions, that will continue processing the other requests if continue on error is true

Related work items: #17661